StateManager is now abstract
You now have to define the behavior of the statemanager. The default
behavior is in the PrincipalMainState class. THis change has been made
to be able to have states having more control on the application than
Abstractstates. For example, the login screen will now be full screen.
References to the persistance "setstate" method Should test if the
statemlanager is an instance of Principalmainstate before calling the
method.
