+++ PMDMMOc/resources/settings.properties
@@ -1,4 +1,4 @@
-#Thu Nov 09 19:49:00 CET 2017
+#Mon Dec 04 00:30:45 CET 2017
 server.address=localhost\:8080/PMDMMOServer/
 key.map.right=39
 key.map.down=40
