+++ PMDMMOc/src/com/darkxell/client/mechanics/freezones/entities/PokemonFreezoneEntity.java
@@ -9,6 +9,7 @@ import com.darkxell.client.mechanics.freezones.FreezoneEntity;
 import com.darkxell.client.resources.images.pokemon.PokemonSprite;
 import com.darkxell.client.state.DialogState;
 import com.darkxell.client.state.DialogState.DialogScreen;
+import com.darkxell.client.state.mainstates.PrincipalMainState;
 import com.darkxell.common.pokemon.Pokemon;
 import com.darkxell.common.pokemon.PokemonRegistry;
 import com.darkxell.common.util.language.Message;
@@ -30,7 +31,8 @@ public class PokemonFreezoneEntity extends FreezoneEntity {
 		screens.add(new DialogScreen(p, new Message("Hey there! <br>This is a debug string to try to know if the dialog boxes are working! Item descriptions are coming next.", false)));
 		screens.add(new DialogScreen(p, new Message("item.info.53")));
 		screens.add(new DialogScreen(PokemonRegistry.find(168).generate(new Random(), 0), new Message("item.info.69")));
-		Persistance.stateManager.setState(new DialogState(Persistance.stateManager.getCurrentState(), screens));
+		if(Persistance.stateManager instanceof PrincipalMainState)
+			((PrincipalMainState) Persistance.stateManager).setState(new DialogState(((PrincipalMainState) Persistance.stateManager).getCurrentState(), screens));
 	}
 
 	@Override
