+++ PMDMMOc/src/com/darkxell/client/state/DialogState.java
@@ -13,6 +13,7 @@ import com.darkxell.client.renderers.TextRenderer.PMDChar;
 import com.darkxell.client.resources.images.MenuHudSpriteset;
 import com.darkxell.client.resources.images.others.Hud;
 import com.darkxell.client.resources.images.pokemon.PokemonPortrait;
+import com.darkxell.client.state.mainstates.PrincipalMainState;
 import com.darkxell.client.ui.Keys;
 import com.darkxell.common.pokemon.Pokemon;
 import com.darkxell.common.util.language.Message;
@@ -148,9 +149,10 @@ public class DialogState extends AbstractState {
 	public void nextMessage() {
 		if (this.currentScreen == this.screens.size() - 1) {
 			if (this.listener == null)
-				Persistance.stateManager.setState(this.backgroundState);
-			else
-				this.listener.onDialogEnd(this);
+				if (Persistance.stateManager instanceof PrincipalMainState)
+					((PrincipalMainState) Persistance.stateManager).setState(this.backgroundState);
+				else
+					this.listener.onDialogEnd(this);
 		} else {
 			++this.currentScreen;
 			this.lines.clear();
