+++ PMDMMOc/src/com/darkxell/client/state/dungeon/ActionSelectionState.java
@@ -9,6 +9,7 @@ import java.awt.image.BufferedImage;
 import com.darkxell.client.launchable.Persistance;
 import com.darkxell.client.resources.images.others.DungeonHudSpriteset;
 import com.darkxell.client.state.dungeon.DungeonState.DungeonSubState;
+import com.darkxell.client.state.mainstates.PrincipalMainState;
 import com.darkxell.client.state.menu.dungeon.DungeonMenuState;
 import com.darkxell.client.ui.Keys;
 import com.darkxell.common.event.TurnSkippedEvent;
@@ -72,7 +73,8 @@ public class ActionSelectionState extends DungeonSubState
 	@Override
 	public void onKeyPressed(short key)
 	{
-		if (key == Keys.KEY_MENU) Persistance.stateManager.setState(new DungeonMenuState(this.parent));
+		if (key == Keys.KEY_MENU) if(Persistance.stateManager instanceof PrincipalMainState)
+			((PrincipalMainState) Persistance.stateManager).setState(new DungeonMenuState(this.parent));
 		if (key == Keys.KEY_ATTACK && (Persistance.player.getDungeonLeader().isFamished() || !Keys.isPressed(Keys.KEY_RUN))) Persistance.eventProcessor
 				.processEvent(new MoveSelectionEvent(Persistance.floor, new LearnedMove(MoveRegistry.ATTACK.id), Persistance.player.getDungeonLeader()));
 	}
