+++ PMDMMOc/src/com/darkxell/client/state/dungeon/NextFloorState.java
@@ -6,6 +6,7 @@ import java.awt.Graphics2D;
 import com.darkxell.client.launchable.Persistance;
 import com.darkxell.client.renderers.TextRenderer;
 import com.darkxell.client.state.AbstractState;
+import com.darkxell.client.state.mainstates.PrincipalMainState;
 import com.darkxell.common.dungeon.Dungeon;
 import com.darkxell.common.util.language.Message;
 
@@ -73,7 +74,8 @@ public class NextFloorState extends AbstractState
 		++this.counter;
 		if (this.counter >= total)
 		{
-			Persistance.stateManager.setState(Persistance.dungeonState);
+			if(Persistance.stateManager instanceof PrincipalMainState)
+				((PrincipalMainState) Persistance.stateManager).setState(Persistance.dungeonState);
 			Persistance.eventProcessor.processPending();
 		}
 		if (this.counter == fade + pause)
