+++ PMDMMOc/src/com/darkxell/client/state/map/DungeonFloorMap.java
@@ -8,6 +8,7 @@ import com.darkxell.client.renderers.floor.PokemonRenderer;
 import com.darkxell.client.resources.images.tilesets.AbstractDungeonTileset;
 import com.darkxell.client.resources.images.tilesets.DungeonMapTileset;
 import com.darkxell.client.state.dungeon.NextFloorState;
+import com.darkxell.client.state.mainstates.PrincipalMainState;
 import com.darkxell.client.ui.Keys;
 import com.darkxell.common.dungeon.floor.Floor;
 import com.darkxell.common.dungeon.floor.Tile;
@@ -103,13 +104,14 @@ public class DungeonFloorMap extends AbstractDisplayMap
 		}
 
 		g.translate(this.x, this.y);
-
-		if (Persistance.stateManager.getCurrentState() instanceof NextFloorState)
-		{
-			int alpha = ((NextFloorState) Persistance.stateManager.getCurrentState()).fading();
-			g.setColor(new Color(0, 0, 0, alpha));
-			g.fillRect(0, 0, width, height);
-		}
+		
+		if(Persistance.stateManager instanceof PrincipalMainState)
+			if (((PrincipalMainState) Persistance.stateManager).getCurrentState() instanceof NextFloorState)
+			{
+				int alpha = ((NextFloorState) ((PrincipalMainState) Persistance.stateManager).getCurrentState()).fading();
+				g.setColor(new Color(0, 0, 0, alpha));
+				g.fillRect(0, 0, width, height);
+			}
 	}
 
 	@Override
