+++ PMDMMOc/src/com/darkxell/client/state/menu/dungeon/DungeonMenuState.java
@@ -7,6 +7,7 @@ import com.darkxell.client.launchable.Persistance;
 import com.darkxell.client.renderers.TextRenderer;
 import com.darkxell.client.state.AbstractState;
 import com.darkxell.client.state.dungeon.DungeonState;
+import com.darkxell.client.state.mainstates.PrincipalMainState;
 import com.darkxell.client.state.menu.OptionSelectionMenuState;
 import com.darkxell.client.state.menu.dungeon.item.ItemContainersMenuState;
 import com.darkxell.common.dungeon.floor.TileType;
@@ -41,14 +42,16 @@ public class DungeonMenuState extends OptionSelectionMenuState
 	@Override
 	protected void onExit()
 	{
-		Persistance.stateManager.setState(this.backgroundState);
+		if(Persistance.stateManager instanceof PrincipalMainState)
+			((PrincipalMainState) Persistance.stateManager).setState(this.backgroundState);
 	}
 
 	@Override
 	protected void onOptionSelected(MenuOption option)
 	{
 		DungeonState s = Persistance.dungeonState;
-		if (option == this.moves) Persistance.stateManager.setState(new MovesMenuState(s));
+		if (option == this.moves) if(Persistance.stateManager instanceof PrincipalMainState)
+			((PrincipalMainState) Persistance.stateManager).setState(new MovesMenuState(s));
 		else if (option == this.items)
 		{
 			ArrayList<ItemContainer> containers = new ArrayList<ItemContainer>();
@@ -60,14 +63,18 @@ public class DungeonMenuState extends OptionSelectionMenuState
 			{
 				this.onExit();
 				s.logger.showMessage(new Message("inventory.empty"));
-			} else Persistance.stateManager.setState(new ItemContainersMenuState(s, containers.toArray(new ItemContainer[containers.size()])));
-		} else if (option == this.team) Persistance.stateManager.setState(new TeamMenuState(s));
+			} else if(Persistance.stateManager instanceof PrincipalMainState)
+				((PrincipalMainState) Persistance.stateManager).setState(new ItemContainersMenuState(s, containers.toArray(new ItemContainer[containers.size()])));
+		} else if (option == this.team) if(Persistance.stateManager instanceof PrincipalMainState)
+			((PrincipalMainState) Persistance.stateManager).setState(new TeamMenuState(s));
 		else if (option == this.ground)
 		{
 			this.onExit();
-			if (Persistance.player.getDungeonLeader().tile.type() == TileType.STAIR) Persistance.stateManager.setState(new StairMenuState());
+			if (Persistance.player.getDungeonLeader().tile.type() == TileType.STAIR) if(Persistance.stateManager instanceof PrincipalMainState)
+				((PrincipalMainState) Persistance.stateManager).setState(new StairMenuState());
 			else if (Persistance.player.getDungeonLeader().tile.getItem() == null) s.logger.showMessage(new Message("ground.empty"));
-			else Persistance.stateManager.setState(new ItemContainersMenuState(s, Persistance.player.getDungeonLeader().tile));
+			else if(Persistance.stateManager instanceof PrincipalMainState)
+				((PrincipalMainState) Persistance.stateManager).setState(new ItemContainersMenuState(s, Persistance.player.getDungeonLeader().tile));
 		}
 	}
